Initial commit
This commit begins the project off with a skeleton implementation
missing many of the planned features. In this version we have a few
usable components in function approximation, the mountain car test
domain and both QLearning and SARSA algorithms. For now the API is very
crude but this will change.

The sandbox.rs test uses an RBFNetwork instance on the mountain car
domain using SARSA with no exploration.