Updated domain observation interface
Added available actions to the observation struct for future use. For
now we don't use this information and no restriction is applied. To make
this work we need to add the appropriate logic to all the policy classes
which will be a nuisance. Really we need a better solution for handling
actions and action spaces since we will eventually want to implement
parameterised actions - which at the moment would be incompatible with
the framework.