Added `pi_target` method to agent trait
This will enable evaluation of an agent using a pure greedy policy.
Beforehand the only interface to sampling an action was via the `pi`
method which interacts with the behaviour (exploration) policy. As such,
there was no way to access the optimal actions given what the agent has
learnt. The `pi_target` method is just that, a method which returns the
optimal action according the agent agent's current knowledge.