Replace pi_target method with evaluate_policy
The pi_target method wasn't well defined for on-policy control
algorithms. It was just being used to evaluate a greedy policy during
evaluation. For more clarity it has been replaced with a method
evaluate_policy which takes a reference to any policy trait object and
evaluates that against the algorithm's internal components. This is
actually more general since it will allow a user construct an agent that
learns an explicitly stochastic policy.