fix(fa): fix `Linear::evaluate_action_phi`
Feature vectors are normalised (using the `l1` norm) in the `Linear`
function approximator to help prevent divergence during learning;
effectively reducing the learning rate. This is especially relevant when
continuous features are used, such as with an `RBFNetwork`, where,
without normalisation, the values may grow very large. It is important
that the application of the norm is uniform across all update and
evaluation routines, otherwise, methods will yield different values for
the same state input.

The bug here was in the `Linear::evaluate_phi` and
`Linear::evaluate_action_phi` methods which were not consistent in their
use of the feature vector. This commit includes a fix which should
prevent issues of this kind.