feat(agents): add `evaluate` method to `PredictionAgent`
We will make the member variables public in future so people can
interact directly with the interval value function representations. With
this one could do their own evaluations.

However, for control algorithms like `ActorCritic`, which use prediction
agents internally, it's helpful to have a generic access point to
evaluating the value of a given state without knowledge of the
underlying implementation.