feat(agents): add `NaturalActorCritic`
This commit adds the Natural Actor-Critic algorithm to the framework. In
order to implement, we add a new trait called `LinearAgent` which may
has one method: `weights`. Intuitively, this method returns the set of
weights used in the internal function approximator instance. This is
implemented in all existing algorithms since one can always implement as
a special case where an instance of Linear has been used.