refactor: replace the QPolicy trait with shared Q functions
In principle, we can avoid the need for a QPolicy trait and many
unimplemented methods in the policies module by storing a reference to
the relevant Q function inside the policies themselves. Ideally one
would just store a reference but this is messy and would require MAJOR
structural changes throughout. The alternative is to store the shared
objects on the heap and use Rc<RefCell<T>>s.

I was initially apprehensive given a pre-formed bias against the heap
for performance reasons. This was a stupid idea (though not unfounded).
It turns out the performance losses are basically negligible.

So... good times!!!

This will allow loads of cool shit - especially given RefCell and Rc
implement CoerceUnsized - and may open this framework up to supporting
multi-agent learning problems!