refactor: major refactor of the core algorithm interfaces
The Algorithm type is now only used as a common trait for the
step_hyperparams method which replaces what was handle_terminal(). We
then introduce two new traits, OnlineLearner and BatchLearner which have
methods specific to the type of learning algorithm. In future it would
be ideal to make these mutually exclusive traits but we don't yet have
this feature in Rust.