refactor(core): rename step_hyperparams -> handle_terminal
There's been a lot of to'ing and fro'ing but I have finally settled on
`handle_terminal`. I think this has a more natural and all encompassing
meaning that includes any updates that may occur to hyperparameters. I
have left it with no arguments to prevent people from doing double
updates. It is expected that people implement their learning code in the
`handle_transition` and/or `handle_batch` methods. The terminal state
function should be used for housekeeping and, maybe, for solving systems
of linear equations as in batch least-squares algos.