feat(core): replace Shared with dedicated type
Before we were using a type alias to Rc<RefCell<T>>. Now we use a
newtype to wrap around the Rc<RefCell<T>> type and implement the
borrow... methods on top. This allows us to implement Deref on top
which, while unsafe, adds a more ergonomic way of calling non-mutable
methods on the interior object. This is based on the gist available at:
https://gist.github.com/stevedonovan/7e3a6d8c8921e3eff16c4b11ab82b8d7.