refactor(policies): reduce mutability requirements for many methods
In a lot of cases the methods for sampling policies or computing the
best action etc required a source of randomness. This was typically done
by having an internal ThreadRng instance. This is ridiculous and made
things very messy, especially with regards to serialisation. Plus, it's
not good practice and makes things more black box which is BAD.

We now provide an instance of rand::Rng to every call of methods that
need it at the call site.