refactor: generalise Gradient code to MatrixLike
This change makes things a little simpler and justifies the methods a
little more (as far as I'm concerned). In future we will want to improve
the linalg code and probably just use components from external crates.
The problem we have is that most existing frameworks don't support
simple versions sparse features or gradient matrices.

My view is to leave it like this for now (i.e. messy AF) and then
rewrite it all to make liberal use of GATs when they are stable. This
would allow us to hide _all_ the underlying code and just implement
lightweight interfaces as with the Entry approach taken in the std
library.