feat: deprecate Parameter type in favour of explicitness and simplicity
This comes after a series of other deprecation commits. The plan is to
do some pruning of code that isn't up to scratch and make the whole
thing simpler and more streamlined. I don't think using a single
`Parameter` type is right (obviously) and instead want to implement a
full suite of hyperparam code; but... there are only so many hours in
the day :'(.