feat(traces): re-write traces code to reduce code duplication
This commit also changes the old Trace trait into a single struct which
is generic with respect to a particular update rule. This is a lot DRYer
and also makes it easier for users to extend.

This also removes the silly use of Deref, and instead just implements
Buffer directly on the Trace struct. You can now pass this directly as a
reference without requiring a call to trace.deref().